+++ backend-app/src/main/java/com/ibcx/poc/service/DepartmentService.java
@@ -8,9 +8,9 @@ import java.util.Optional;
 public interface DepartmentService {
 
 	public List<Department> getDepartments();
-	public Optional<Department> getDepartmentById(int departmentId);
+	public Optional<Department> getDepartmentById(Long departmentId);
 	public Department addNewDepartment(Department department) throws Exception;
 	public Department updateDepartment(Department department) throws Exception;
-	public void deleteDepartmentById(int departmentId) throws Exception;
+	public void deleteDepartmentById(Long departmentId) throws Exception;
 
 }
